# Add same_value casting to np.astype