--- a/numpy/_core/include/numpy/dtype_api.h
+++ b/numpy/_core/include/numpy/dtype_api.h
@@ -107,13 +107,15 @@ typedef struct PyArrayMethod_Context_tag {
 
     /* Operand descriptors, filled in by resolve_descriptors */
     PyArray_Descr *const *descriptors;
-    void * padding;
+ #if NPY_FEATURE_VERSION > NPY_2_3_API_VERSION
+    void * _reserved;
     /* 
      * Optional flag to pass information into the inner loop
      * If set, it will be NPY_CASTING
      */
     uint64_t flags;
     /* Structure may grow (this is harmless for DType authors) */
+ #endif
 } PyArrayMethod_Context;
 
 
@@ -149,12 +151,6 @@ typedef struct {
 #define NPY_METH_unaligned_contiguous_loop 8
 #define NPY_METH_contiguous_indexed_loop 9
 #define _NPY_METH_static_data 10
-
-/*
- * Constants for same_value casting
- */
-#define NPY_SAME_VALUE_OVERFLOW -31
-
 
 /*
  * The resolve descriptors function, must be able to handle NULL values for