--- a/numpy/_core/src/multiarray/array_assign_scalar.c
+++ b/numpy/_core/src/multiarray/array_assign_scalar.c
@@ -37,7 +37,7 @@
 NPY_NO_EXPORT int
 raw_array_assign_scalar(int ndim, npy_intp const *shape,
         PyArray_Descr *dst_dtype, char *dst_data, npy_intp const *dst_strides,
-        PyArray_Descr *src_dtype, char *src_data)
+        PyArray_Descr *src_dtype, char *src_data, NPY_CASTING casting)
 {
     int idim;
     npy_intp shape_it[NPY_MAXDIMS], dst_strides_it[NPY_MAXDIMS];
@@ -85,6 +85,11 @@ raw_array_assign_scalar(int ndim, npy_intp const *shape,
         }
         NPY_BEGIN_THREADS_THRESHOLDED(nitems);
     }
+    if (casting == NPY_SAME_VALUE_CASTING) {
+        cast_info.context.flags |= NPY_SAME_VALUE_CASTING;
+        PyErr_SetString(PyExc_NotImplementedError, "'same_value' casting not implemented yet");
+        return -1;
+    }
 
     npy_intp strides[2] = {0, dst_strides_it[0]};
 
@@ -126,7 +131,7 @@ raw_array_wheremasked_assign_scalar(int ndim, npy_intp const *shape,
         PyArray_Descr *dst_dtype, char *dst_data, npy_intp const *dst_strides,
         PyArray_Descr *src_dtype, char *src_data,
         PyArray_Descr *wheremask_dtype, char *wheremask_data,
-        npy_intp const *wheremask_strides)
+        npy_intp const *wheremask_strides, NPY_CASTING casting)
 {
     int idim;
     npy_intp shape_it[NPY_MAXDIMS], dst_strides_it[NPY_MAXDIMS];
@@ -176,6 +181,11 @@ raw_array_wheremasked_assign_scalar(int ndim, npy_intp const *shape,
             nitems *= shape_it[i];
         }
         NPY_BEGIN_THREADS_THRESHOLDED(nitems);
+    }
+    if (casting == NPY_SAME_VALUE_CASTING) {
+        cast_info.context.flags |= NPY_SAME_VALUE_CASTING;
+        PyErr_SetString(PyExc_NotImplementedError, "'same_value' casting not implemented yet");
+        return -1;
     }
 
     npy_intp strides[2] = {0, dst_strides_it[0]};
@@ -298,7 +308,7 @@ PyArray_AssignRawScalar(PyArrayObject *dst,
         /* Do the assignment with raw array iteration */
         if (raw_array_assign_scalar(PyArray_NDIM(dst), PyArray_DIMS(dst),
                 PyArray_DESCR(dst), PyArray_DATA(dst), PyArray_STRIDES(dst),
-                src_dtype, src_data) < 0) {
+                src_dtype, src_data, casting) < 0) {
             goto fail;
         }
     }
@@ -319,7 +329,7 @@ PyArray_AssignRawScalar(PyArrayObject *dst,
                 PyArray_DESCR(dst), PyArray_DATA(dst), PyArray_STRIDES(dst),
                 src_dtype, src_data,
                 PyArray_DESCR(wheremask), PyArray_DATA(wheremask),
-                wheremask_strides) < 0) {
+                wheremask_strides, casting) < 0) {
             goto fail;
         }
     }