--- a/numpy/_core/src/multiarray/array_method.c
+++ b/numpy/_core/src/multiarray/array_method.c
@@ -190,6 +190,7 @@ validate_spec(PyArrayMethod_Spec *spec)
         case NPY_SAFE_CASTING:
         case NPY_SAME_KIND_CASTING:
         case NPY_UNSAFE_CASTING:
+        case NPY_SAME_VALUE_CASTING:
             break;
         default:
             if (spec->casting != -1) {