… doc/source/reference/c-api/array.rst
… numpy/_core/code_generators/cversions.txt
… numpy/_core/src/multiarray/array_method.h
… numpy/_core/src/multiarray/common.c
… numpy/_core/src/multiarray/conversion_utils.c
… numpy/_core/src/multiarray/convert.c
… numpy/_core/src/multiarray/convert_datatype.c
… numpy/_core/src/multiarray/ctors.c
… numpy/_core/src/multiarray/datetime.c
… numpy/_core/src/multiarray/datetime_strings.c
… numpy/_core/src/multiarray/dtype_transfer.c
… numpy/_core/src/multiarray/dtype_transfer.h
… numpy/_core/src/multiarray/einsum.c.src
… numpy/_core/src/multiarray/legacy_dtype_implementation.c
… numpy/_core/src/multiarray/lowlevel_strided_loops.c.src
… numpy/_core/src/multiarray/mapping.c
… numpy/_core/src/multiarray/methods.c
… numpy/_core/src/multiarray/multiarraymodule.c
… numpy/_core/src/umath/legacy_array_method.c
… numpy/_core/src/umath/reduction.c
… numpy/_core/src/umath/ufunc_object.c
… numpy/_core/tests/test_casting_unittests.py
… numpy/_core/tests/test_datetime.py
… numpy/_core/tests/test_einsum.py
… numpy/_core/tests/test_shape_base.py