This is something I have wanted for a while, to allow casting to a "smaller" dtype and raise an error or warning if there is overflow. For now this is a PR to my branch, so we can have a place to discuss the progress without adding too much noise to the general repo.