@mattip 2025-05-27
Choices for implementing the same_value cast: 
- A dedicated inner loop, which would double the number of casting loops: we already have 4: aligned/not aligned and strided/contiguous, so that would mean 8 loop functions
code around the cast in the inner loop and a if flag==SAMEVALUE branch. 
- Adding a flags field to the context. This will impact performance 

Checking only the flag on the fast path 
```
    if (context->flags != 1234) {
        *(_TYPE2 *)dst = _CONVERT_FN(*(_TYPE1 *)src);
    }

```

makes this timeit test go from 640ns to 880 ns:
```
timeit -s "import numpy as np; a = np.arange(10_000, dtype=np.int64)" "a.astype(np.int8)"
```

Increasing the size to 10_000_000 and testing with/without the dummy flag makes the time go from 74.5ms to 69.5ms. The casting loops already use `NPY_GCC_OPT_3`.