@mattip 2025-06-30
I have made this into a PR to numpy/numpy